pyproject.toml
[project]
name = "holonai-hackathon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dotenv>=0.9.9",
    "langchain>=0.3.23",
    "langgraph>=0.3.29",
    "streamlit>=1.44.1",
    "pandas>=2.2.3",
    "watchdog>=6.0.0",
    "langchain-google-genai>=2.1.2",
    "langgraph-reflection>=0.0.1",
    "plotly>=6.0.1",
    "google-cloud-bigquery>=3.31.0",
]
